--- botV2/cogBot/database.py
@@ -1,10 +1,0 @@
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-from sqlalchemy.ext.declarative import declarative_base
-
-SQLALCHEMY_DATABASE_URI = 'sqlite:///clients.sqlite'
-
-engine = create_engine(SQLALCHEMY_DATABASE_URI, connect_args={'check_same_thread': False})
-SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
-Base = declarative_base()
