--- botV2/cogBot/models.py
@@ -1,33 +1,0 @@
-from sqlalchemy import Column, Integer, String
-from sqlalchemy.sql.expression import null
-from sqlalchemy.sql.sqltypes import Boolean
-from database import Base
-
-class Clients(Base):
-    __tablename__ = 'clients'
-    id = Column(Integer, primary_key=True, index=True)
-    guild_id = Column(Integer, index=True, nullable=False)
-    notion_api_key = Column(String, nullable=False)
-    notion_db_id = Column(String, nullable=False)
-    tag = Column(Boolean, default=False)
-    contributor = Column(Boolean, default=False)
-    prefix = Column(String, default="!")
-
-    def __init__(self, guild_id, notion_api_key, notion_db_id, tag, contributor, prefix="!"):
-        self.guild_id = guild_id
-        self.notion_api_key = notion_api_key
-        self.notion_db_id = notion_db_id
-        self.tag = tag
-        self.prefix = prefix
-        self.contributor = contributor
-
-    @property
-    def serialize(self):
-        return {
-            "guild_id": self.guild_id,
-            "notion_api_key": self.notion_api_key,
-            "notion_db_id": self.notion_db_id,
-            "tag": self.tag,
-            "prefix": self.prefix,
-            "contributor": self.contributor
-        }
