+++ botV2/cogs/add.py
