+++ botV2/cogs/delete.py
