+++ botV2/functionality/addRecord.py
