+++ botV2/functionality/deleteRecord.py
