+++ botV2/functionality/setupBot.py
