+++ botV2/functionality/utils.py
